+++ src/flask/blueprints.py
@@ -344,13 +344,15 @@ class Blueprint(Scaffold):
                 app.cli.add_command(self.cli)
 
         for blueprint, bp_options in self._blueprints:
-            url_prefix = options.get("url_prefix", "")
+            bp_url_prefix = state.url_prefix or ""
             if "url_prefix" in bp_options:
-                url_prefix = (
-                    url_prefix.rstrip("/") + "/" + bp_options["url_prefix"].lstrip("/")
+                bp_url_prefix = (
+                    bp_url_prefix.rstrip("/")
+                    + "/"
+                    + bp_options["url_prefix"].lstrip("/")
                 )
 
-            bp_options["url_prefix"] = url_prefix
+            bp_options["url_prefix"] = bp_url_prefix
             bp_options["name_prefix"] = options.get("name_prefix", "") + self.name + "."
             blueprint.register(app, bp_options)